Add files via upload
This is my personal portfolio website designed to showcase my projects, skills, and experiences as a budding Electronics and Communication Engineer. Built using HTML, CSS, and JavaScript, the site features a project gallery that highlights my work and accomplishments, as well as a contact form for visitors to reach out to me directly. 

 I plan to enhance the website further by adding a blog section to share insights and tutorials, and improving its responsiveness for a better mobile experience.